feat(KFLUXINFRA-741): Add "Local" cloud
The "Local" cloud configuration is meant to allow a remote-capable
piepline task to run locally in the cluster by creating a dummy SSH
secret for it that only has the "host" key set to "localhost.

While we could in theory create a host pool that contains "localhost",
that would've caused the controller to try to SSH into it to configure
user accounts, etc.