fix(KFLUXINFRA-836): Bump tekton API version
Using version 0.61.0 of the Tekton Go client API in order to be
compatible with Tekton of that version and above.

To do this had to Bump our Go version to 1.22.7 as well as a bunch of
other dependencies.

Since we build on a UBI Go toolset image that only has Go 1.21.x,
setting the `GOTOOLSET` variable to `auto`  during our container
builds to pull the proper toolset version as part of the build.

This commit is much large then it needs to be because of the vendored
dependencies.