Fix backend probe counters (#429)
* metrics: avoid potentially ignored atomic writes

This code is subtly incorrect, since it has the potential for some
success or failure events to be ignored.

The root of this race is because we do loads and stores on both the
success and failure counters in two separate atomic operations, rather
than 1 each.  This means that any modification of the two variables
between their loads and the deferred stores will be lost.

To avoid this race, perform both a load and a store in a single atomic
operation on each of these variables, which gives no opportunity for a
modification of these counters to be missed.

Furthermore, rather than re-loading these variables multiple times, only
load them once at the start of this method and reuse their loaded values
during checks.

Signed-off-by: Andy Sadler <ansadler@redhat.com>

* metrics: change backend probe counters to unsigned

For these counters, negative values are semantically meaningless and
should be impossible to reach.  Represent them as unsigned int32s
instead.

Signed-off-by: Andy Sadler <ansadler@redhat.com>

---------

Signed-off-by: Andy Sadler <ansadler@redhat.com>