Tag new VMs and clean up orphaned VMs (AWS & Power)
This commit is part of a feature to create a clean up mechanism to
delete orphaned VMs (i.e. those associated with a non-existent TaskRun).
AWS EC2 VMs and IBM Power Cloud PVMs are tagged with their TaskRun's
namespace and name on creation and both cloud providers implement a
clean up function that iterates over all VM instances, retrieves the
TaskRun tag, and removes any VMs not found to be associated with an
existing TaskRun.