Add functionality to re-queue failed IBM VMs (#425)
Since MPC assumes a cloud provider's capacity is limitless, it will
create a virtual machine (VM) on the cloud provider and if the cloud
provider cannot start the VM due to a lack of capacity (which is
indicated by the VM being in a failed state), the user's build will fail.
This commit checks if an IBM VM has entered a failed state and re-queues
it after 1 minute if it has.