Tag new VMs and clean up orphaned VMs (IBM Z)
This commit is part of a feature to create a clean up mechanism to
delete orphaned VMs (i.e. those associated with a non-existent TaskRun).
Volumes of IBM Virtual Private Cloud VMs are tagged with their TaskRun's
namespace and name on creation and both cloud providers implement a
clean up function that iterates over all VM instances, retrieves the
volume, retrieves the TaskRun tag, and removes any VMs not found to be
associated with an existing TaskRun. VM volumes are tagged because the
VPC API does not allow user tags on VM instances.