taskrun: isolate schemas between tests
When running tests concurrently (ginkgo run --race -p ./...), go would
intermittently complain about races accessing a clients scheme object.
This is happening because we always try to add Kubernetes and Tekton
schemas to the scheme builder, which is racy when we run tests
concurrently.

To fix this, construct new schemas for each test to isolate unit tests
from each other.

Signed-off-by: Andy Sadler <ansadler@redhat.com>