test: enable concurrency
Many of our tests don't depend on each other, so we should be able to
run tests concurrently.  Add the `-p` flag to ginkgo to enable this.

For ginkgo tests, we can always add the `Ordered` flag to test setups to
run tests sequentially.

As long as we're enabling concurrent tests, let's also enable the race
detector to catch concurrency issues.

Signed-off-by: Andy Sadler <ansadler@redhat.com>