Tag new VMs and clean up orphaned VMs (#439)
* Tag new VMs and clean up orphaned VMs (AWS & Power)

This commit is part of a feature to create a clean up mechanism to
delete orphaned VMs (i.e. those associated with a non-existent TaskRun).
AWS EC2 VMs and IBM Power Cloud PVMs are tagged with their TaskRun's
namespace and name on creation and both cloud providers implement a
clean up function that iterates over all VM instances, retrieves the
TaskRun tag, and removes any VMs not found to be associated with an
existing TaskRun.

* Tag new VMs and clean up orphaned VMs (IBM VPC)

This commit is part of a feature to create a clean up mechanism to
delete orphaned VMs (i.e. those associated with a non-existent TaskRun).
IBM VPC volumes are tagged with their TaskRun's namespace and name on
creation of their associated VM. The IBM VPC cloud provider implements a
clean up function that iterates over all VM instances, gets their
associated volume, retrieves the TaskRun tag, and removes any VMs not
found to be associated with an existing TaskRun.

* Adding unit tests for CleanUpVMs functionality

This commit adds unit tests to verify the underlying functionality of
the methods used in the AWS EC2 and IBM Power Cloud cloud providers'
`CleanUpVMs` function. These unit tests should test the ability to
determine if an instance is associated with a non-existent TaskRun.