ci: use a better golangci-lint setup (#546)
* ci: upload sarif report from linting job

Github can render linting output as feedback on PRs.  To enable this, a
sarif-formatted report needs to be uploaded using the
`github/codeql-action/upload-sarif` action.  Since golangci-lint
supports generating these sarif manifests, enable them in MPC's CI jobs.

Part-of: KFLUXINFRA-2265
Signed-off-by: Andy Sadler <ansadler@redhat.com>

* ci: use golanci-lint version from modfile

The version of golangci-lint run locally is not necessarily the same
version as what's run in CI.  This can cause issues when the two are out
of sync.

To remedy this, pin the version of golangci-lint to the version
specified in the modfile.

Part-of: KFLUXINFRA-2265
Signed-off-by: Andy Sadler <ansadler@redhat.com>

---------

Signed-off-by: Andy Sadler <ansadler@redhat.com>