ci: bump golangci-lint to v2.4.0 (#547)
* bump golangci-lint to v2.4.0

We've been running an out-of-date version of golangci-lint for a while
now, so let's bump it.

Part-of: KFLUXINFRA-2265
Signed-off-by: Andy Sadler <ansadler@redhat.com>

* add golangci-lint configuration file

If we want to add more linters or formatters in the future beyond the
defaults, we'll need a configuration file to add them to.

Part-of: KFLUXINFRA-2265
Signed-off-by: Andy Sadler <ansadler@redhat.com>

* fix linting errors

Fix the outstanding linter warning recognized by the new linter version.

Part-of: KFLUXINFRA-2265
Signed-off-by: Andy Sadler <ansadler@redhat.com>

---------

Signed-off-by: Andy Sadler <ansadler@redhat.com>