Use golangci-lint for formatting (#558)
* run formatters through golangci-lint

golangci-lint has more formatters available, so let's use it for both
linting and formatting going forward.

Part-of: KFLUXINFRA-2265
Signed-off-by: Andy Sadler <ansadler@redhat.com>

* fmt: add `goimports` formatter

Ensure that imports across all our modules are always sorted, and that
we have a consistent formatting style across all modules.

Part-of: KFLUXINFRA-2265
Signed-off-by: Andy Sadler <ansadler@redhat.com>

* ci: use diff from fmt output

golangci-lint can give us the diff directly, so we don't have to compute
it ourselves.  Add the necessary flags for diff output.

Part-of: KFLUXINFRA-2265
Signed-off-by: Andy Sadler <ansadler@redhat.com>

---------

Signed-off-by: Andy Sadler <ansadler@redhat.com>